+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.1.2 [unreleased]
 * improvements on the README and stack example - thanks @Boran
+* read device scanner settings using --query_properties
 
 ## 3.1.1
 * add support for ScaleIO 2.0.1